# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ConsumerApp API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/client/flat": {
            "get": {
                "tags": [
                    "ClientFlat"
                ],
                "description": "Получение списка связок клиент-квартира",
                "responses": {
                    "200": {
                        "description": "Список связок клиент-квартира найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Массив найденных связок клиент-квартира",
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ClientFlat"
                                                    }
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ClientFlat"
                ],
                "description": "Добавить связку клиент-квартира",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientFlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Связка клиент-квартира добавлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/ClientFlat"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/flat/{flat}": {
            "get": {
                "tags": [
                    "ClientFlat"
                ],
                "description": "Получение связки клиент-квартира",
                "responses": {
                    "200": {
                        "description": "Связка клиент-квартира найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/ClientFlat"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ClientFlat"
                ],
                "description": "Отредактировать связку клиент-квартира",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientFlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Связка клиент-квартира отредактирована",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Client"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ClientFlat"
                ],
                "description": "Удалить связоку клиент-квартира",
                "responses": {
                    "200": {
                        "description": "Связка клиент-квартира удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "null"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ClientFlatRequest": {
                "properties": {
                    "client_id": {
                        "description": "Идентификатор клиента",
                        "type": "integer"
                    },
                    "flat_id": {
                        "description": "Идентификатор квартиры",
                        "type": "integer"
                    },
                    "client_flat_status_id": {
                        "description": "Идентификатор статуса связки клиент-квартира",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "ClientFlat": {
                "properties": {
                    "id": {
                        "description": "Идентификатор связки клиент-квартира",
                        "type": "integer"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "flat": {
                        "$ref": "#/components/schemas/Flat"
                    },
                    "client_flat_status": {
                        "$ref": "#/components/schemas/ClientFlatStatus"
                    }
                },
                "type": "object"
            },
            "ClientFlatStatus": {
                "properties": {
                    "id": {
                        "description": "Идентификатор статуса связки клиент-квартира",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название статуса связки клиент-квартира",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Client": {
                "properties": {
                    "id": {
                        "description": "Идентификатор клиента",
                        "type": "integer"
                    },
                    "name": {
                        "description": "ФИО клиента",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "email": {
                        "description": "Почта клиента",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Телефон клиента",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Flat": {
                "properties": {
                    "id": {
                        "description": "Идентификатор квартиры",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название квартиры",
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/FlatStatus"
                    },
                    "full_space": {
                        "description": "Вся площадь",
                        "type": "number",
                        "format": "float"
                    },
                    "floor_count": {
                        "description": "Кол-во этажей",
                        "type": "integer"
                    },
                    "living_space": {
                        "description": "Жилая площадь",
                        "type": "number",
                        "format": "float"
                    },
                    "room_count": {
                        "description": "Кол-во комнат",
                        "type": "integer"
                    },
                    "balconyless_space": {
                        "description": "Площадьь без балконов и т.п.",
                        "type": "number",
                        "format": "float"
                    },
                    "residential_complex_id": {
                        "description": "Идентификатор ЖК, в котором располагается квартира",
                        "type": "integer"
                    },
                    "cost": {
                        "description": "Начальная цена квартиры",
                        "type": "integer"
                    },
                    "is_ready": {
                        "description": "Готова ли квартира",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "FlatStatus": {
                "properties": {
                    "id": {
                        "description": "Идентификатор статуса квартиры",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название статуса квартиры",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer"
                    },
                    "email": {
                        "description": "Почта пользователя",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Когда пользователь был зарегистрирован",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}